Fix: size_t overflow, barrier correctness, named constant
- max_output_segments_ changed from int to size_t (removes truncating cast)
- Cast all operands in max_segs multiplication to size_t
- Add GL_BUFFER_UPDATE_BARRIER_BIT before CPU readback of max SSBO
- Replace magic 32 with named kLineSegBytes constant

+++ src/core/renderer.h
@@ -76,7 +76,7 @@ private:
     // Output segment SSBO + draw indirect buffer
     GLuint output_ssbo_ = 0;
     GLuint draw_cmd_buffer_ = 0;
-    int max_output_segments_ = 0;
+    size_t max_output_segments_ = 0;
 
     // Trace shader uniform locations
     GLint trace_loc_num_circles_ = -1;
